requirements.txt
# Web Framework
Flask==2.0.1
Flask-CORS==3.0.10
Werkzeug==2.0.1
gunicorn==20.1.0

# Database
SQLAlchemy==1.4.23
Flask-SQLAlchemy==2.5.1
Flask-Login==0.5.0
pymongo==3.12.0
flask-pymongo==2.3.0

# Authentication & Security
PyJWT==2.1.0
bcrypt==3.2.0
python-dotenv==0.19.0

# External APIs
spotipy==2.19.0
google-api-python-client==2.15.0
google-generativeai==0.1.0
newsapi-python==0.2.6
requests==2.26.0

# File Handling
python-magic==0.4.24
mutagen==1.45.1
fpdf==1.7.2

# Data Processing
beautifulsoup4==4.9.3
feedparser==6.0.8
lxml==4.9.0

# Blockchain
web3==5.31.1
eth-account==0.5.9

# Testing
pytest==6.2.5
pytest-cov==2.12.1

# Utilities
python-dateutil==2.8.2
Pillow==8.3.2

# Bluetooth & IoT
bleak==0.21.1
asyncio==3.4.3
aiohttp==3.8.5
pandas==1.5.3

# Real-time Data Processing
numpy==1.24.3
scikit-learn==1.3.0
